requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
astor==0.8.1
asyncpg==0.31.0
attrs==26.1.0
bce-python-sdk==0.9.59
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
datamodel-code-generator==0.53.0
decorator==5.2.1
dotenv==0.9.9
easyocr==1.7.2
fastapi==0.128.0
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
genson==1.3.0
google_search_results==2.4.2
greenlet==3.3.0
h11==0.16.0
hdbscan==0.8.41
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
inflect==7.5.0
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.8
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
open_clip_torch==3.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle-gpu==2.6.1
pandas==2.3.3
pathspec==1.0.4
pgvector==0.4.2
pillow==12.0.0
platformdirs==4.5.1
prettytable==3.17.0
propcache==0.4.1
protobuf==3.20.2
psutil==7.2.1
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.3.1
pypdfium2==5.3.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.5
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
selenium==4.41.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.25
tokenizers==0.22.2
tomli==2.4.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.6
trio==0.33.0
trio-websocket==0.12.2
typeguard==4.4.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.23.0
